Refactor borrowing system and introduce settings GUI
- Cleaned up whitespace and formatting in borrowing_system.py for improved readability.
- Introduced a new settings_gui.py module to handle user settings configuration before the main application starts.
- Removed unused imports and refactored Movemental.py to streamline the codebase.
- Enhanced the BorrowingController class with improved state management and visual feedback for borrowing operations.
- Updated borrowing logic to ensure better handling of active notes and chord display updates.
